# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZeeoWork API Name",
        "description": "ZeeoWork API Description",
        "contact": {
            "email": "tech@zeework.ir"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://devpandora.zeework.ir/"
        }
    ],
    "paths": {
...
        "/api/admin/profile": {
            "get": {
                "tags": [
                    "Admin/Profile"
                ],
                "summary": "Get current admin profile",
                "operationId": "ba377a553b379f7d4f9896fb6c79c130",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin/Profile"
                ],
                "summary": "Update current admin profile",
                "operationId": "408e0d8007fb33e0f30cfc4438596fd0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAdminProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/profile/change-password": {
            "put": {
                "tags": [
                    "Admin/Profile"
                ],
                "summary": "Change current admin password",
                "operationId": "8a4a264fade37b48402e2a430a635063",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeAdminPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/employee/organizations": {
            "get": {
                "tags": [
                    "Employee/Organizations"
                ],
                "summary": "Get the list of organizations for the logged-in employee",
                "operationId": "9a821d82b7cf3c935e626cb00494c143",
                "responses": {
                    "200": {
                        "description": "A list of the employee's organizations.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdminOrganizationResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/employee/profile": {
            "get": {
                "tags": [
                    "Employee/Profile"
                ],
                "summary": "Get current Employee profile",
                "operationId": "3a5192fe5690ff714fba58f0371e7d9b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeSelfResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Employee/Profile"
                ],
                "summary": "Update current Employee profile",
                "operationId": "9c6977eabcca49ed420e833eeb0048cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEmployeeProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/employee/requests": {
            "get": {
                "tags": [
                    "Employee/Requests"
                ],
                "summary": "Get list of health requests for the logged-in employee",
                "operationId": "90e4d5afd5cf41fbd688e06c139522c5",
                "parameters": [
                    {
                        "name": "organization_id",
                        "in": "query",
                        "description": "فیلتر بر اساس سازمان",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "فیلتر بر اساس وضعیت",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EmployeeRequestListResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error: organization_id is required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employee/Requests"
                ],
                "summary": "Create a new personal health request by an Employee",
                "operationId": "6e809148d5695c3447ec74add4c31a1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeStorePersonalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EmployeeRequestResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/employee/requests/{id}": {
            "get": {
                "tags": [
                    "Employee/Requests"
                ],
                "summary": "Get details of a specific health request (Task Dashboard)",
                "operationId": "d1e3dd4aa6bcfcf096d78a72abdd482f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeRequestDetailsResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Request not found or access denied"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "messages": {
                        "description": "آرایه‌ای از پیام‌ها برای کاربر (می‌تواند شامل پیام‌های موفقیت، خطا یا اطلاعاتی باشد)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "عملیات با موفقیت انجام شد."
                        },
                        "example": [
                            "عملیات موفقیت‌آمیز بود",
                            "پیام دوم در صورت وجود"
                        ],
                        "nullable": true
                    },
                    "data": {
                        "type": "object",
                        "example": {
                            "id": 1,
                            "name": "Ali"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ChangeAdminPasswordRequest": {
                "title": "Change Admin Password Request",
                "required": [
                    "current_password",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "current_password": {
                        "description": "Admin's current password",
                        "type": "string",
                        "format": "password",
                        "example": "current_secret"
                    },
                    "password": {
                        "description": "Admin's new password",
                        "type": "string",
                        "format": "password",
                        "example": "NewS3cr3t!"
                    },
                    "password_confirmation": {
                        "description": "Confirmation of the new password",
                        "type": "string",
                        "format": "password",
                        "example": "NewS3cr3t!"
                    }
                },
                "type": "object"
            },
...
            "UpdateAdminProfileRequest": {
                "title": "Update Admin Profile Request",
                "required": [],
                "properties": {
                    "name": {
                        "description": "Admin's first name",
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "description": "Admin's email",
                        "type": "string",
                        "example": "Doe@test.com"
                    },
                    "phone": {
                        "description": "Admin's phone",
                        "type": "string",
                        "example": "09012211111"
                    }
                },
                "type": "object"
            },
...
            "EmployeeStorePersonalRequest": {
                "title": "Employee Store Personal Request",
                "required": [
                    "organization_id"
                ],
                "properties": {
                    "organization_id": {
                        "description": "ID سازمانی که درخواست برای آن ثبت می‌شود",
                        "type": "integer",
                        "example": 1
                    },
                    "options": {
                        "properties": {
                            "selected_scores": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "diabetes_risk"
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UpdateEmployeeProfileRequest": {
                "title": "Update Employee Profile Request",
                "description": "Request body for updating employee profile information. Email and phone number changes might be ignored or handled separately.",
                "required": [],
                "properties": {
                    "name": {
                        "description": "Employee's full name or first name",
                        "type": "string",
                        "example": "پوریا"
                    },
                    "national_code": {
                        "description": "Employee's national code (if submitted, must be unique if changed)",
                        "type": "string",
                        "example": "0012345678"
                    },
                    "email": {
                        "description": "Employee's email (if submitted, must be unique if changed)",
                        "type": "string",
                        "format": "email",
                        "example": "employee@example.com"
                    },
                    "phone": {
                        "description": "Employee's phone number (if submitted, must be unique if changed)",
                        "type": "string",
                        "example": "09123456789"
                    },
                    "photo": {
                        "description": "Employee's photo/avatar image file (optional)",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "AdminResource": {
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
            "AdminOrganizationResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "name": {
                        "type": "string",
                        "example": "شرکت فناوری آرمان"
                    },
                    "national_id": {
                        "type": "string",
                        "example": "14001234567",
                        "nullable": true
                    },
                    "reg_number": {
                        "type": "string",
                        "example": "123456",
                        "nullable": true
                    },
                    "economic_code": {
                        "type": "string",
                        "example": "987654321",
                        "nullable": true
                    },
                    "logo": {
                        "type": "string",
                        "format": "url",
                        "example": "https://cdn.example.com/logos/5.png",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "تهران، خیابان آزادی، پلاک ۲۳",
                        "nullable": true
                    },
                    "company_phone": {
                        "type": "string",
                        "example": "02112345678",
                        "nullable": true
                    },
                    "representative_name": {
                        "type": "string",
                        "example": "رضا محمدی",
                        "nullable": true
                    },
                    "representative_position": {
                        "type": "string",
                        "example": "مدیرعامل",
                        "nullable": true
                    },
                    "representative_phone": {
                        "type": "string",
                        "example": "09121234567",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
            "EmployeeSelfResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "national_code": {
                        "type": "string",
                        "example": "0071234567"
                    },
                    "name": {
                        "type": "string",
                        "example": "علی رضایی"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ali.rezaei@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "09123456789"
                    },
                    "photo": {
                        "type": "string",
                        "example": "https://example.com/uploads/employees/123.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
            "EmployeeRequestListResource": {
                "title": "Employee Request List Item Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "tracking_code": {
                        "type": "string",
                        "example": "14030408-42",
                        "nullable": true
                    },
                    "status": {
                        "description": "وضعیت کلی درخواست",
                        "type": "string",
                        "example": "pending"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-29T10:00:00Z"
                    },
                    "organization": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "شرکت راهبران"
                            },
                            "logo": {
                                "type": "string",
                                "format": "uri",
                                "example": "https://example.com/logo.png"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "EmployeeRequestResource": {
                "title": "Employee Request Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tracking_code": {
                        "type": "string",
                        "example": "14030408-1"
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "in_process"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-28T12:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}